+++ src/app/_hooks/channel/write/useFilePreview.ts
@@ -1,0 +1,15 @@
+import { useState } from 'react';
+
+export default function useFilePreview() {
+  const [selectedFile, setSelectedFile] = useState<File | null>(null);
+  const [previewUrl, setPreviewUrl] = useState("");
+
+  const onFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+    if (!file) return;
+    setSelectedFile(file);
+    setPreviewUrl(URL.createObjectURL(file));
+  };
+
+  return { selectedFile, previewUrl, onFileChange };
+}
